+++ hetzner-ubuntu18-zfs-setup.sh
@@ -732,35 +732,37 @@ for ((i = 1; i < ${#v_selected_disks[@]}; i++)); do
   dd if="${v_selected_disks[0]}-part1" of="${v_selected_disks[i]}-part1"
 done
 
-echo "=========set up dropbear=============="
-chroot_execute "apt install --yes dropbear-initramfs"
+if [[ $v_encrypt_rpool == "1" ]]; then
+  echo "=========set up dropbear=============="
+  chroot_execute "apt install --yes dropbear-initramfs"
 
-cp /root/.ssh/authorized_keys "$c_zfs_mount_dir/etc/dropbear-initramfs/authorized_keys"
+  cp /root/.ssh/authorized_keys "$c_zfs_mount_dir/etc/dropbear-initramfs/authorized_keys"
 
-cp "$c_zfs_mount_dir/etc/ssh/ssh_host_rsa_key" "$c_zfs_mount_dir/etc/ssh/ssh_host_rsa_key_temp"
-chroot_execute "ssh-keygen -p -i -m pem -N '' -f /etc/ssh/ssh_host_rsa_key_temp"
-chroot_execute "/usr/lib/dropbear/dropbearconvert openssh dropbear /etc/ssh/ssh_host_rsa_key_temp /etc/dropbear-initramfs/dropbear_rsa_host_key"
-rm -rf "$c_zfs_mount_dir/etc/ssh/ssh_host_rsa_key_temp"
+  cp "$c_zfs_mount_dir/etc/ssh/ssh_host_rsa_key" "$c_zfs_mount_dir/etc/ssh/ssh_host_rsa_key_temp"
+  chroot_execute "ssh-keygen -p -i -m pem -N '' -f /etc/ssh/ssh_host_rsa_key_temp"
+  chroot_execute "/usr/lib/dropbear/dropbearconvert openssh dropbear /etc/ssh/ssh_host_rsa_key_temp /etc/dropbear-initramfs/dropbear_rsa_host_key"
+  rm -rf "$c_zfs_mount_dir/etc/ssh/ssh_host_rsa_key_temp"
 
-cp "$c_zfs_mount_dir/etc/ssh/ssh_host_ecdsa_key" "$c_zfs_mount_dir/etc/ssh/ssh_host_ecdsa_key_temp"
-chroot_execute "ssh-keygen -p -i -m pem -N '' -f /etc/ssh/ssh_host_ecdsa_key_temp"
-chroot_execute "/usr/lib/dropbear/dropbearconvert openssh dropbear /etc/ssh/ssh_host_ecdsa_key_temp /etc/dropbear-initramfs/dropbear_ecdsa_host_key"
-chroot_execute "rm -rf /etc/ssh/ssh_host_ecdsa_key_temp"
-rm -rf "$c_zfs_mount_dir/etc/ssh/ssh_host_ecdsa_key_temp"
+  cp "$c_zfs_mount_dir/etc/ssh/ssh_host_ecdsa_key" "$c_zfs_mount_dir/etc/ssh/ssh_host_ecdsa_key_temp"
+  chroot_execute "ssh-keygen -p -i -m pem -N '' -f /etc/ssh/ssh_host_ecdsa_key_temp"
+  chroot_execute "/usr/lib/dropbear/dropbearconvert openssh dropbear /etc/ssh/ssh_host_ecdsa_key_temp /etc/dropbear-initramfs/dropbear_ecdsa_host_key"
+  chroot_execute "rm -rf /etc/ssh/ssh_host_ecdsa_key_temp"
+  rm -rf "$c_zfs_mount_dir/etc/ssh/ssh_host_ecdsa_key_temp"
 
-rm -rf "$c_zfs_mount_dir/etc/dropbear-initramfs/dropbear_dss_host_key"
+  rm -rf "$c_zfs_mount_dir/etc/dropbear-initramfs/dropbear_dss_host_key"
 
-cd "$c_zfs_mount_dir/root"
-wget http://ftp.de.debian.org/debian/pool/main/libt/libtommath/libtommath1_1.1.0-3_amd64.deb
-wget http://ftp.de.debian.org/debian/pool/main/d/dropbear/dropbear-bin_2018.76-5_amd64.deb
-wget http://ftp.de.debian.org/debian/pool/main/d/dropbear/dropbear-initramfs_2018.76-5_all.deb
+  cd "$c_zfs_mount_dir/root"
+  wget http://ftp.de.debian.org/debian/pool/main/libt/libtommath/libtommath1_1.1.0-3_amd64.deb
+  wget http://ftp.de.debian.org/debian/pool/main/d/dropbear/dropbear-bin_2018.76-5_amd64.deb
+  wget http://ftp.de.debian.org/debian/pool/main/d/dropbear/dropbear-initramfs_2018.76-5_all.deb
 
-chroot_execute "dpkg -i /root/libtommath1_1.1.0-3_amd64.deb"
-chroot_execute "dpkg -i /root/dropbear-bin_2018.76-5_amd64.deb"
-chroot_execute "dpkg -i /root/dropbear-initramfs_2018.76-5_all.deb"
+  chroot_execute "dpkg -i /root/libtommath1_1.1.0-3_amd64.deb"
+  chroot_execute "dpkg -i /root/dropbear-bin_2018.76-5_amd64.deb"
+  chroot_execute "dpkg -i /root/dropbear-initramfs_2018.76-5_all.deb"
 
-rm $c_zfs_mount_dir/root/*.deb
-cd /root
+  rm $c_zfs_mount_dir/root/*.deb
+  cd /root
+fi
 
 echo "============setup root prompt============"
 cat > "$c_zfs_mount_dir/root/.bashrc" <<CONF
